remove old/unused transactions route code
