+++ src/routes/safe/components/Transactions/NoTransactions/index.jsx
@@ -5,8 +5,10 @@ import Col from '~/components/layout/Col'
 import Row from '~/components/layout/Row'
 import Paragraph from '~/components/layout/Paragraph/index'
 
-const NoRights = () => (
-  <Row>
+export const NO_TRANSACTION_ROW_TEST_ID = 'no-transaction-row'
+
+const NoTransactions = () => (
+  <Row data-testid={NO_TRANSACTION_ROW_TEST_ID}>
     <Col xs={12} center="xs" sm={10} smOffset={2} start="sm" margin="md">
       <Paragraph size="lg">
         <Bold>No transactions found for this safe</Bold>
@@ -15,4 +17,4 @@ const NoRights = () => (
   </Row>
 )
 
-export default NoRights
+export default NoTransactions
