+++ src/routes/safe/components/Transactions/TxsTable/ExpandedTx/ApproveTxModal/style.js
